requirements.txt
pyyaml~=5.3.1
requests~=2.24.0
bs4~=0.0.1
beautifulsoup4~=4.9.1
python-docx